requirements.txt
Django>=1.4.0
django-tastypie>=0.9.11
